UPSKILL WEATHER PROJECT
This project is basically a system designed and programmed with raspberry pi( a computer itself)
and a sensor called SENSE HAT used to detect and measure temperature and humidity readings from an
environment. The readings are displayed on the screen, stored in a file and are sent as notifications to a smartphone. Stay tuned!